LayoutTests/imported/w3c/web-platform-tests/css/css-scroll-anchoring/ancestor-change-heuristic-expected.txt


FAIL Ancestor changes in document scroller. assert_equals: expected 150 but got 220
FAIL Ancestor changes in scrollable <div>. assert_equals: expected 200 but got 150
